Keep base TextAttributes as part of AttributedString (#46612)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/46612

This information is lost by the time the AttributedString makes its way to be measured.

Pretty sure I've seen code do incorrect heuristics like grabbing first span for this (IIRC adjustFontSizeToFit does this)? Another case where we hit this on Android is when the last character is a newline (which is part of the Spannable), and layout needs to predict the size of the line which will come next (where it uses the EditText's underlying TextPaint, who is using these base attributes).

Changelog: [Internal]

Reviewed By: sammy-SC

Differential Revision: D63303709

fbshipit-source-id: 644a97c80a9280e643514786ea6cf32faa4f1cf8