Add an event start timestamp field into RawEvent (#46949)
Summary:

# Changelog:

[Internal] - 

This adds an (optional) event start time field into RawEvent, so that it can be threaded by the platform-specific code to tell where is the earliest point of the event start (depending on the platform).

For example, on Android platform it can be the point of time when the original MotionEvent was received.

Differential Revision: D64175467