Remove implicit "start", then clean up callsites (#47313)
Summary:

We're going to fix up a bunch of designated initializers.  `RCTSurfaceHostingProxyRootView` is particularly problematic because different initializers will do different things even though reading the code it looks like they should be equivalent.  Remove the encapsulated "start" to the provided "surface", and require it to be explicit at the callsite.

Reviewed By: cipolleschi

Differential Revision: D65214656